requirements.txt
opencv-python
pytesseract
PyQt5
Pillow

